Fix widgets e2e tests integration to core
