+++ packages/e2e-tests/plugins/classic-widgets.php
@@ -1,0 +1,10 @@
+<?php
+/**
+ * Plugin Name: Gutenberg Test Classic Widgets
+ * Plugin URI: https://github.com/WordPress/gutenberg
+ * Author: Gutenberg Team
+ *
+ * @package gutenberg-test-classic-widgets
+ */
+
+add_filter( 'use_widgets_block_editor', '__return_false' );
