+++ packages/e2e-tests/specs/widgets/customizing-widgets.test.js
@@ -16,6 +16,11 @@ import {
 // eslint-disable-next-line no-restricted-imports
 import { find } from 'puppeteer-testing-library';
 
+/**
+ * Internal dependencies
+ */
+import { cleanupWidgets } from './test-utils';
+
 describe( 'Widgets Customizer', () => {
 	beforeEach( async () => {
 		await cleanupWidgets();
@@ -511,28 +516,6 @@ describe( 'Widgets Customizer', () => {
 		);
 	} );
 } );
-
-/**
- * TODO: Deleting widgets in the new widgets screen seems to be unreliable.
- * We visit the old widgets screen to delete them.
- * Refactor this to use real interactions in the new widgets screen once the bug is fixed.
- */
-async function cleanupWidgets() {
-	await visitAdminPage( 'widgets.php' );
-
-	let widget = await page.$( '.widgets-sortables .widget' );
-
-	// We have to do this one-by-one since there might be race condition when deleting multiple widgets at once.
-	while ( widget ) {
-		const deleteButton = await widget.$( 'button.widget-control-remove' );
-		const id = await widget.evaluate( ( node ) => node.id );
-		await deleteButton.evaluate( ( node ) => node.click() );
-		// Wait for the widget to be removed from DOM.
-		await page.waitForSelector( `#${ id }`, { hidden: true } );
-
-		widget = await page.$( '.widgets-sortables .widget' );
-	}
-}
 
 /**
  * Wait when there's only one preview iframe.
